Fix SEXP->TREE -- Double-quoted TREE-TYPE args
The quote->comma only worked when no symbol was passed for the TREE-TYPE
parameter, otherwise %SEXP->TREE would wind up with a TREE-TYPE argument
that was double-quoted.

To fix this, the second quote was moved to the default value of the
TREE-TYPE parameter of SEXP->TREE, thus %SEXP->TREE always gets a
single-quoted TREE-TYPE argument.